+++ lib/decorators/use.decorator.ts
@@ -1,6 +1,5 @@
 import { Request, Response } from 'express';
 import { Api } from '../templates/api';
-import { Middleware } from '../templates/middleware';
 
 export const USE = Symbol('__use__');
 
@@ -11,10 +10,10 @@ export type MiddlewareFn = (
 ) => void;
 
 export interface UseMetadata {
-  middleware: (new () => Middleware) | MiddlewareFn;
+  middleware: MiddlewareFn;
 }
 
-export function Use(middleware: (new () => Middleware) | MiddlewareFn) {
+export function Use(middleware: MiddlewareFn) {
   return function (target: new () => Api<any, any, any>) {
     Reflect.defineMetadata(USE, { middleware } as UseMetadata, target);
   };