+++ src/modules/categories/apis/get-categories.api.ts
@@ -1,8 +1,8 @@
-import { Api, Get, Module } from '../../../../lib';
+import { Api, HttpGet, Module } from '../../../../lib';
 import { Category } from '../entities/category.entity';
 
 @Module('categories')
-@Get('all')
+@HttpGet('all')
 export class GetCatoriesApi extends Api {
   private readonly categoryRep = this.db.getRepository(Category);