+++ src/modules/users/apis/get-user.api.ts
@@ -1,9 +1,9 @@
-import { Api, Module, Get, Priority } from '../../../../lib';
+import { Api, Module, HttpGet, Priority } from '../../../../lib';
 import { GetUserDto } from '../dto/get-user.dto';
 import { User } from '../entities/user.entity';
 
 @Module('users')
-@Get(':user_id')
+@HttpGet(':user_id')
 @Priority(1)
 export class GetUserApi extends Api<GetUserDto> {
   userRep = this.db.getRepository(User);