+++ test/fixtures/create.api.ts
@@ -1,5 +1,5 @@
 import { IsString } from 'class-validator';
-import { Api, Body, Module, NotFoundError, Post } from '../../lib';
+import { Api, Body, Module, NotFoundError, HttpPost } from '../../lib';
 
 export class CreateBodyDto {
   @IsString()
@@ -7,7 +7,7 @@ export class CreateBodyDto {
 }
 
 @Module('clientes')
-@Post('')
+@HttpPost('')
 @Body(CreateBodyDto)
 export class CreateCustomerApi extends Api<null, CreateBodyDto> {
   main() {
@@ -17,7 +17,7 @@ export class CreateCustomerApi extends Api<null, CreateBodyDto> {
 
 /** Throws a domain error: it must come out mapped to 404 by ErrorControl. */
 @Module('clientes')
-@Post('falla')
+@HttpPost('falla')
 export class FailingApi extends Api {
   main(): never {
     throw new NotFoundError('cliente no existe');